refactor: move conversation projection into harness core
Replaces the split JSON-RPC/session snapshot projection logic with a shared harness-owned reducer and projected-item contract, moves model-stream normalization into shared code, and switches desktop live hydration onto projected turn/item snapshots instead of provider-specific re-projection.

Updates the JSON-RPC schema/docs/generated artifacts, covers repeated raw ids plus non-turn projected items and ask/approval parity in tests, and records the verification trail in tasks/todo.md.

Verification:
- bun test test/jsonrpc.projectors.test.ts test/sessionSnapshotProjector.test.ts test/jsonrpc.thread-read-projector.test.ts apps/desktop/test/control-socket.test.ts apps/desktop/test/protocol-v2-events.test.ts test/server.jsonrpc.flow.test.ts
- bun run typecheck
- bun run docs:generate-jsonrpc
- bun test
