+++ apps/desktop/src/app/modelStream.ts
@@ -2,14 +2,14 @@ export type {
   ModelStreamChunkEvent,
   ModelStreamRawEvent,
   ModelStreamUpdate,
-} from "@cowork/client/modelStream";
+} from "../../../../src/shared/modelStream";
 export type {
   ModelStreamReplayRuntime,
-} from "@cowork/client/modelStreamReplay";
-export { mapModelStreamChunk, mapModelStreamRawEvent } from "@cowork/client/modelStream";
+} from "../../../../src/shared/modelStreamReplay";
+export { mapModelStreamChunk, mapModelStreamRawEvent } from "../../../../src/shared/modelStream";
 export {
   clearModelStreamReplayRuntime,
   createModelStreamReplayRuntime,
   replayModelStreamRawEvent,
   shouldIgnoreNormalizedChunkForRawBackedTurn,
-} from "@cowork/client/modelStreamReplay";
+} from "../../../../src/shared/modelStreamReplay";
