+++ src/shared/modelStream.ts
@@ -1,0 +1,576 @@
+import { z } from "zod";
+
+import type { ServerEvent } from "../server/protocol";
+
+const stringSchema = z.string();
+const finiteNumberSchema = z.number().finite();
+const partRecordSchema = z.record(z.string(), z.unknown());
+
+export type ModelStreamChunkEvent = Extract<ServerEvent, { type: "model_stream_chunk" }>;
+export type ModelStreamRawEvent = Extract<ServerEvent, { type: "model_stream_raw" }>;
+
+export type ModelStreamUpdate =
+  | { kind: "turn_start"; turnId: string }
+  | {
+      kind: "turn_finish";
+      turnId: string;
+      finishReason?: unknown;
+      rawFinishReason?: unknown;
+      totalUsage?: unknown;
+    }
+  | { kind: "turn_abort"; turnId: string; reason?: unknown }
+  | { kind: "turn_error"; turnId: string; error: unknown }
+  | { kind: "step_start"; turnId: string; stepNumber?: number; request?: unknown; warnings?: unknown }
+  | {
+      kind: "step_finish";
+      turnId: string;
+      stepNumber?: number;
+      response?: unknown;
+      usage?: unknown;
+      finishReason?: unknown;
+      rawFinishReason?: unknown;
+      providerMetadata?: unknown;
+    }
+  | { kind: "assistant_text_start"; turnId: string; streamId: string; phase?: string }
+  | {
+      kind: "assistant_delta";
+      turnId: string;
+      streamId: string;
+      text: string;
+      phase?: string;
+    }
+  | {
+      kind: "assistant_text_end";
+      turnId: string;
+      streamId: string;
+      phase?: string;
+      annotations?: Array<Record<string, unknown>>;
+    }
+  | { kind: "reasoning_start"; turnId: string; streamId: string; mode: "reasoning" | "summary" }
+  | { kind: "reasoning_delta"; turnId: string; streamId: string; mode: "reasoning" | "summary"; text: string }
+  | { kind: "reasoning_end"; turnId: string; streamId: string; mode: "reasoning" | "summary" }
+  | { kind: "tool_input_start"; turnId: string; key: string; name: string; args?: unknown }
+  | { kind: "tool_input_delta"; turnId: string; key: string; delta: string }
+  | { kind: "tool_input_end"; turnId: string; key: string; name: string }
+  | { kind: "tool_call"; turnId: string; key: string; name: string; args?: unknown }
+  | { kind: "tool_result"; turnId: string; key: string; name: string; result: unknown }
+  | { kind: "tool_error"; turnId: string; key: string; name: string; error: unknown }
+  | { kind: "tool_output_denied"; turnId: string; key: string; name: string; reason?: unknown }
+  | { kind: "tool_approval_request"; turnId: string; approvalId: string; toolCall: unknown }
+  | { kind: "source"; turnId: string; source: unknown }
+  | { kind: "file"; turnId: string; file: unknown }
+  | { kind: "raw"; turnId: string; raw: unknown }
+  | { kind: "unknown"; turnId: string; partType: string; payload: unknown };
+
+function asString(value: unknown): string | undefined {
+  const parsed = stringSchema.safeParse(value);
+  return parsed.success ? parsed.data : undefined;
+}
+
+function asIdString(value: unknown): string | undefined {
+  if (typeof value === "string") {
+    const trimmed = value.trim();
+    return trimmed.length > 0 ? trimmed : undefined;
+  }
+  if (typeof value === "number") {
+    return Number.isFinite(value) ? String(value) : undefined;
+  }
+  if (typeof value === "bigint") {
+    return value.toString();
+  }
+  return undefined;
+}
+
+function asLooseText(value: unknown): string | undefined {
+  if (typeof value === "string") return value;
+  if (typeof value === "number") return Number.isFinite(value) ? String(value) : undefined;
+  if (typeof value === "boolean") return String(value);
+  if (typeof value === "bigint") return value.toString();
+  if (value && typeof value === "object") {
+    try {
+      return JSON.stringify(value);
+    } catch {
+      return undefined;
+    }
+  }
+  return undefined;
+}
+
+function asFiniteNumber(value: unknown): number | undefined {
+  const parsed = finiteNumberSchema.safeParse(value);
+  return parsed.success ? parsed.data : undefined;
+}
+
+function asPartRecord(value: unknown): Record<string, unknown> {
+  const parsed = partRecordSchema.safeParse(value);
+  return parsed.success ? parsed.data : {};
+}
+
+function asRecordArray(value: unknown): Array<Record<string, unknown>> | undefined {
+  if (!Array.isArray(value)) return undefined;
+  const next = value.filter(
+    (entry): entry is Record<string, unknown> =>
+      typeof entry === "object" && entry !== null && !Array.isArray(entry),
+  );
+  return next.length > 0 ? next : undefined;
+}
+
+function asReasoningMode(value: unknown): "reasoning" | "summary" {
+  return value === "summary" ? "summary" : "reasoning";
+}
+
+function toolKey(evt: ModelStreamChunkEvent): string {
+  const part = asPartRecord(evt.part);
+  return (
+    asIdString(part.toolCallId) ??
+    asIdString(part.id) ??
+    `tool:${evt.turnId}:${evt.index}`
+  );
+}
+
+function toolName(evt: ModelStreamChunkEvent): string {
+  const part = asPartRecord(evt.part);
+  return asString(part.toolName) ?? "tool";
+}
+
+function rawProviderKey(part: Record<string, unknown>, fallback: string): string {
+  return (
+    asIdString(part.item_id) ??
+    asIdString(part.itemId) ??
+    asIdString(part.call_id) ??
+    asIdString(part.callId) ??
+    asIdString(part.tool_call_id) ??
+    asIdString(part.toolCallId) ??
+    asIdString(part.function_call_id) ??
+    asIdString(part.functionCallId) ??
+    asIdString(part.invocation_id) ??
+    asIdString(part.invocationId) ??
+    asIdString(part.id) ??
+    fallback
+  );
+}
+
+function assistantPhase(value: unknown): string | undefined {
+  const phase = asString(value)?.trim();
+  return phase ? phase : undefined;
+}
+
+function mapProviderStreamEvent(
+  evt: ModelStreamChunkEvent,
+  eventType: string,
+  payload: Record<string, unknown>
+): ModelStreamUpdate | null {
+  const normalizedType = eventType.toLowerCase();
+
+  if (normalizedType.includes("function_call_arguments") && normalizedType.endsWith(".delta")) {
+    const delta = asLooseText(payload.delta) ?? asLooseText(payload.arguments);
+    if (delta === undefined) return null;
+    return {
+      kind: "tool_input_delta",
+      turnId: evt.turnId,
+      key: rawProviderKey(payload, `raw-tool:${evt.turnId}:${evt.index}`),
+      delta,
+    };
+  }
+
+  if (
+    normalizedType.includes("function_call_arguments") &&
+    (normalizedType.endsWith(".done") || normalizedType.endsWith(".completed"))
+  ) {
+    return {
+      kind: "tool_input_end",
+      turnId: evt.turnId,
+      key: rawProviderKey(payload, `raw-tool:${evt.turnId}:${evt.index}`),
+      name: asString(payload.name) ?? asString(payload.tool_name) ?? "tool",
+    };
+  }
+
+  if (normalizedType.includes("reasoning") && normalizedType.endsWith(".delta")) {
+    const text =
+      asLooseText(payload.delta) ??
+      asLooseText(payload.text) ??
+      asLooseText(payload.summary) ??
+      asLooseText(payload.content);
+    if (text === undefined) return null;
+    return {
+      kind: "reasoning_delta",
+      turnId: evt.turnId,
+      streamId: rawProviderKey(payload, `raw-reasoning:${evt.turnId}:${evt.index}`),
+      mode: normalizedType.includes("summary") ? "summary" : "reasoning",
+      text,
+    };
+  }
+
+  if (
+    normalizedType.includes("reasoning") &&
+    (normalizedType.endsWith(".done") || normalizedType.endsWith(".completed"))
+  ) {
+    return {
+      kind: "reasoning_end",
+      turnId: evt.turnId,
+      streamId: rawProviderKey(payload, `raw-reasoning:${evt.turnId}:${evt.index}`),
+      mode: normalizedType.includes("summary") ? "summary" : "reasoning",
+    };
+  }
+
+  if (
+    normalizedType.includes("reasoning") &&
+    (normalizedType.endsWith(".start") || normalizedType.endsWith(".started"))
+  ) {
+    return {
+      kind: "reasoning_start",
+      turnId: evt.turnId,
+      streamId: rawProviderKey(payload, `raw-reasoning:${evt.turnId}:${evt.index}`),
+      mode: normalizedType.includes("summary") ? "summary" : "reasoning",
+    };
+  }
+
+  if (normalizedType.includes("output_text") && normalizedType.endsWith(".delta")) {
+    const text = asLooseText(payload.delta) ?? asLooseText(payload.text);
+    if (text === undefined) return null;
+    return {
+      kind: "assistant_delta",
+      turnId: evt.turnId,
+      streamId: rawProviderKey(payload, `raw-text:${evt.turnId}:${evt.index}`),
+      text,
+      ...(assistantPhase(payload.phase) ? { phase: assistantPhase(payload.phase) } : {}),
+    };
+  }
+
+  if (
+    normalizedType.includes("output_text") &&
+    (normalizedType.endsWith(".done") || normalizedType.endsWith(".completed"))
+  ) {
+    return {
+      kind: "assistant_text_end",
+      turnId: evt.turnId,
+      streamId: rawProviderKey(payload, `raw-text:${evt.turnId}:${evt.index}`),
+      ...(assistantPhase(payload.phase) ? { phase: assistantPhase(payload.phase) } : {}),
+    };
+  }
+
+  if (
+    normalizedType.includes("output_text") &&
+    (normalizedType.endsWith(".start") || normalizedType.endsWith(".started"))
+  ) {
+    return {
+      kind: "assistant_text_start",
+      turnId: evt.turnId,
+      streamId: rawProviderKey(payload, `raw-text:${evt.turnId}:${evt.index}`),
+      ...(assistantPhase(payload.phase) ? { phase: assistantPhase(payload.phase) } : {}),
+    };
+  }
+
+  if (normalizedType === "response.completed") {
+    const response = asPartRecord(payload.response);
+    return {
+      kind: "turn_finish",
+      turnId: evt.turnId,
+      finishReason: response.status ?? payload.status ?? "completed",
+      rawFinishReason: response.error ?? payload.error,
+      totalUsage: response.usage ?? payload.usage,
+    };
+  }
+
+  if (normalizedType === "response.failed") {
+    return {
+      kind: "turn_error",
+      turnId: evt.turnId,
+      error: payload.error ?? payload,
+    };
+  }
+
+  return null;
+}
+
+export function mapModelStreamChunk(evt: ModelStreamChunkEvent): ModelStreamUpdate | null {
+  const part = asPartRecord(evt.part);
+  switch (evt.partType) {
+    case "start":
+      return { kind: "turn_start", turnId: evt.turnId };
+    case "finish":
+      return {
+        kind: "turn_finish",
+        turnId: evt.turnId,
+        finishReason: part.finishReason,
+        rawFinishReason: part.rawFinishReason,
+        totalUsage: part.totalUsage,
+      };
+    case "abort":
+      return {
+        kind: "turn_abort",
+        turnId: evt.turnId,
+        reason: part.reason,
+      };
+    case "error":
+      return {
+        kind: "turn_error",
+        turnId: evt.turnId,
+        error: part.error ?? "unknown_error",
+      };
+    case "start_step": {
+      const startStepNumber = asFiniteNumber(part.stepNumber);
+      return {
+        kind: "step_start",
+        turnId: evt.turnId,
+        ...(startStepNumber !== undefined ? { stepNumber: startStepNumber } : {}),
+        request: part.request,
+        warnings: part.warnings,
+      };
+    }
+    case "finish_step": {
+      const finishStepNumber = asFiniteNumber(part.stepNumber);
+      return {
+        kind: "step_finish",
+        turnId: evt.turnId,
+        ...(finishStepNumber !== undefined ? { stepNumber: finishStepNumber } : {}),
+        response: part.response,
+        usage: part.usage,
+        finishReason: part.finishReason,
+        rawFinishReason: part.rawFinishReason,
+        providerMetadata: part.providerMetadata,
+      };
+    }
+    case "text_start":
+      return {
+        kind: "assistant_text_start",
+        turnId: evt.turnId,
+        streamId: asString(part.id) ?? `text:${evt.index}`,
+        ...(assistantPhase(part.phase) ? { phase: assistantPhase(part.phase) } : {}),
+      };
+    case "text_delta": {
+      const text = asString(part.text);
+      if (text === undefined) {
+        return {
+          kind: "unknown",
+          turnId: evt.turnId,
+          partType: "text_delta",
+          payload: part,
+        };
+      }
+      return {
+        kind: "assistant_delta",
+        turnId: evt.turnId,
+        streamId: asString(part.id) ?? `text:${evt.index}`,
+        text,
+        ...(assistantPhase(part.phase) ? { phase: assistantPhase(part.phase) } : {}),
+      };
+    }
+    case "text_end":
+      return {
+        kind: "assistant_text_end",
+        turnId: evt.turnId,
+        streamId: asString(part.id) ?? `text:${evt.index}`,
+        ...(asRecordArray(part.annotations) ? { annotations: asRecordArray(part.annotations) } : {}),
+        ...(assistantPhase(part.phase) ? { phase: assistantPhase(part.phase) } : {}),
+      };
+    case "reasoning_start":
+      return {
+        kind: "reasoning_start",
+        turnId: evt.turnId,
+        streamId: asString(part.id) ?? `reasoning:${evt.index}`,
+        mode: asReasoningMode(part.mode),
+      };
+    case "reasoning_delta": {
+      const text = asLooseText(part.text);
+      if (text === undefined) {
+        return {
+          kind: "unknown",
+          turnId: evt.turnId,
+          partType: "reasoning_delta",
+          payload: part,
+        };
+      }
+      return {
+        kind: "reasoning_delta",
+        turnId: evt.turnId,
+        streamId: asString(part.id) ?? `reasoning:${evt.index}`,
+        mode: asReasoningMode(part.mode),
+        text,
+      };
+    }
+    case "reasoning_end":
+      return {
+        kind: "reasoning_end",
+        turnId: evt.turnId,
+        streamId: asString(part.id) ?? `reasoning:${evt.index}`,
+        mode: asReasoningMode(part.mode),
+      };
+    case "tool_input_start":
+      return {
+        kind: "tool_input_start",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        name: toolName(evt),
+        args: part,
+      };
+    case "tool_input_delta": {
+      const delta = asLooseText(part.delta);
+      if (delta === undefined) {
+        return {
+          kind: "unknown",
+          turnId: evt.turnId,
+          partType: "tool_input_delta",
+          payload: part,
+        };
+      }
+      return {
+        kind: "tool_input_delta",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        delta,
+      };
+    }
+    case "tool_input_end":
+      return {
+        kind: "tool_input_end",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        name: toolName(evt),
+      };
+    case "tool_call":
+      return {
+        kind: "tool_call",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        name: toolName(evt),
+        args: part.input,
+      };
+    case "tool_result":
+      return {
+        kind: "tool_result",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        name: toolName(evt),
+        result: part.output,
+      };
+    case "tool_error":
+      return {
+        kind: "tool_error",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        name: toolName(evt),
+        error: part.error,
+      };
+    case "tool_output_denied":
+      return {
+        kind: "tool_output_denied",
+        turnId: evt.turnId,
+        key: toolKey(evt),
+        name: toolName(evt),
+        reason: part.reason,
+      };
+    case "tool_approval_request":
+      return {
+        kind: "tool_approval_request",
+        turnId: evt.turnId,
+        approvalId: asString(part.approvalId) ?? `${evt.turnId}:${evt.index}`,
+        toolCall: part.toolCall,
+      };
+    case "source":
+      return {
+        kind: "source",
+        turnId: evt.turnId,
+        source: part.source,
+      };
+    case "file":
+      return {
+        kind: "file",
+        turnId: evt.turnId,
+        file: part.file,
+      };
+    case "raw":
+      {
+        const rawPayload = asPartRecord(part.raw);
+        const fallbackRaw = Object.keys(rawPayload).length > 0 ? rawPayload : asPartRecord(evt.rawPart);
+        const payload = Object.keys(fallbackRaw).length > 0 ? fallbackRaw : rawPayload;
+        const rawType = asString(payload.type) ?? asString(payload.event_type);
+        const rawMapped = rawType ? mapProviderStreamEvent(evt, rawType, payload) : null;
+        if (rawMapped) return rawMapped;
+      }
+      return {
+        kind: "raw",
+        turnId: evt.turnId,
+        raw: part.raw,
+      };
+    case "unknown":
+      {
+        const sdkType = asString(part.sdkType) ?? asString(part.type);
+        const rawPayload = asPartRecord(part.raw);
+        const fallbackRaw = Object.keys(rawPayload).length > 0 ? rawPayload : asPartRecord(evt.rawPart);
+        const payload = Object.keys(fallbackRaw).length > 0 ? fallbackRaw : rawPayload;
+        const fallbackSdkType = asString(payload.type) ?? asString(payload.event_type);
+        const unknownMapped = (sdkType ?? fallbackSdkType)
+          ? mapProviderStreamEvent(evt, sdkType ?? (fallbackSdkType as string), payload)
+          : null;
+        if (unknownMapped) return unknownMapped;
+      }
+      return {
+        kind: "unknown",
+        turnId: evt.turnId,
+        partType: "unknown",
+        payload: part,
+      };
+    default:
+      return {
+        kind: "unknown",
+        turnId: evt.turnId,
+        partType: String(evt.partType),
+        payload: part,
+      };
+  }
+}
+
+function webSearchCallAction(item: Record<string, unknown>): Record<string, unknown> | undefined {
+  const action = asPartRecord(item.action);
+  return Object.keys(action).length > 0 ? action : undefined;
+}
+
+function webSearchCallSources(item: Record<string, unknown>): unknown[] | undefined {
+  const action = webSearchCallAction(item);
+  if (!action || !Array.isArray(action.sources)) return undefined;
+  return action.sources;
+}
+
+function mapNativeWebSearchRawEvent(evt: ModelStreamRawEvent): ModelStreamUpdate[] {
+  if (evt.format !== "openai-responses-v1") return [];
+
+  const event = asPartRecord(evt.event);
+  const eventType = asString(event.type);
+  if (!eventType) return [];
+
+  const item = asPartRecord(event.item);
+  if (item.type !== "web_search_call") return [];
+
+  const key = rawProviderKey(item, `native-web-search:${evt.turnId}:${evt.index}`);
+  if (eventType === "response.output_item.added") {
+    return [{
+      kind: "tool_input_start",
+      turnId: evt.turnId,
+      key,
+      name: "nativeWebSearch",
+      args: webSearchCallAction(item),
+    }];
+  }
+
+  if (eventType === "response.output_item.done") {
+    return [{
+      kind: "tool_result",
+      turnId: evt.turnId,
+      key,
+      name: "nativeWebSearch",
+      result: {
+        status: item.status ?? "completed",
+        action: webSearchCallAction(item),
+        sources: webSearchCallSources(item),
+        raw: item,
+      },
+    }];
+  }
+
+  return [];
+}
+
+export function mapModelStreamRawEvent(evt: ModelStreamRawEvent): ModelStreamUpdate[] {
+  return mapNativeWebSearchRawEvent(evt);
+}
